comint: save & restore fonts so misidentified prompts don't lose their color
I'm hitting a very annoying bug where sometimes colors will be missing
from random segments of output, always starting at the beginning of a
line.

The reason is that comint is seeing a partial line, interpreting it as
a prompt, coloring it as a prompt, and then uncoloring it when it
receives more input.

This should fix it by saving and restoring the colors in an alternate
text property instead.